# contracts/openapi/trusted-publication.v1.yaml
openapi: 3.0.3
info:
  title: NeKiro Trusted Publication API
  version: 1.0.0
paths:
  /v1/providers/{providerId}/agents/{agentId}/releases:
    post:
      operationId: createAgentRelease
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/providerId' }
        - { $ref: '#/components/parameters/agentId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAgentReleaseRequest' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/AgentRelease' } } } }
        '400': { $ref: '#/components/responses/TrustedPublicationError' }
        '401': { $ref: '#/components/responses/TrustedPublicationError' }
        '403': { $ref: '#/components/responses/TrustedPublicationError' }
        '404': { $ref: '#/components/responses/TrustedPublicationError' }
        '409': { $ref: '#/components/responses/TrustedPublicationError' }
        '503': { $ref: '#/components/responses/TrustedPublicationError' }
        '500': { $ref: '#/components/responses/TrustedPublicationError' }
  /v1/releases/{releaseId}:
    get:
      operationId: getAgentRelease
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/releaseId' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AgentRelease' } } } }
        '401': { $ref: '#/components/responses/TrustedPublicationError' }
        '403': { $ref: '#/components/responses/TrustedPublicationError' }
        '404': { $ref: '#/components/responses/TrustedPublicationError' }
        '503': { $ref: '#/components/responses/TrustedPublicationError' }
        '500': { $ref: '#/components/responses/TrustedPublicationError' }
  /v1/releases/{releaseId}/verify:
    post:
      operationId: verifyAgentRelease
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/releaseId' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AgentRelease' } } } }
        '401': { $ref: '#/components/responses/TrustedPublicationError' }
        '403': { $ref: '#/components/responses/TrustedPublicationError' }
        '404': { $ref: '#/components/responses/TrustedPublicationError' }
        '409': { $ref: '#/components/responses/TrustedPublicationError' }
        '503': { $ref: '#/components/responses/TrustedPublicationError' }
        '500': { $ref: '#/components/responses/TrustedPublicationError' }
  /v1/releases/{releaseId}/publish:
    post:
      operationId: publishAgentRelease
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/releaseId' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AgentRelease' } } } }
        '401': { $ref: '#/components/responses/TrustedPublicationError' }
        '403': { $ref: '#/components/responses/TrustedPublicationError' }
        '404': { $ref: '#/components/responses/TrustedPublicationError' }
        '409': { $ref: '#/components/responses/TrustedPublicationError' }
        '503': { $ref: '#/components/responses/TrustedPublicationError' }
        '500': { $ref: '#/components/responses/TrustedPublicationError' }
  /v1/releases/{releaseId}/suspend:
    post:
      operationId: suspendAgentRelease
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/releaseId' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AgentRelease' } } } }
        '401': { $ref: '#/components/responses/TrustedPublicationError' }
        '403': { $ref: '#/components/responses/TrustedPublicationError' }
        '404': { $ref: '#/components/responses/TrustedPublicationError' }
        '409': { $ref: '#/components/responses/TrustedPublicationError' }
        '503': { $ref: '#/components/responses/TrustedPublicationError' }
        '500': { $ref: '#/components/responses/TrustedPublicationError' }
  /v1/releases/{releaseId}/revoke:
    post:
      operationId: revokeAgentRelease
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/releaseId' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AgentRelease' } } } }
        '401': { $ref: '#/components/responses/TrustedPublicationError' }
        '403': { $ref: '#/components/responses/TrustedPublicationError' }
        '404': { $ref: '#/components/responses/TrustedPublicationError' }
        '409': { $ref: '#/components/responses/TrustedPublicationError' }
        '503': { $ref: '#/components/responses/TrustedPublicationError' }
        '500': { $ref: '#/components/responses/TrustedPublicationError' }
  /v1/providers/{providerId}/agents/{agentId}/endpoint-bindings:
    post:
      operationId: createEndpointBinding
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/providerId' }
        - { $ref: '#/components/parameters/agentId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateEndpointBindingRequest' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/EndpointBinding' } } } }
        '400': { $ref: '#/components/responses/TrustedPublicationError' }
        '401': { $ref: '#/components/responses/TrustedPublicationError' }
        '403': { $ref: '#/components/responses/TrustedPublicationError' }
        '404': { $ref: '#/components/responses/TrustedPublicationError' }
        '409': { $ref: '#/components/responses/TrustedPublicationError' }
        '503': { $ref: '#/components/responses/TrustedPublicationError' }
        '500': { $ref: '#/components/responses/TrustedPublicationError' }
  /v1/providers/{providerId}/endpoint-bindings/{bindingId}:
    get:
      operationId: getEndpointBinding
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/providerId' }
        - { $ref: '#/components/parameters/bindingId' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/EndpointBinding' } } } }
        '401': { $ref: '#/components/responses/TrustedPublicationError' }
        '403': { $ref: '#/components/responses/TrustedPublicationError' }
        '404': { $ref: '#/components/responses/TrustedPublicationError' }
        '400': { $ref: '#/components/responses/TrustedPublicationError' }
        '503': { $ref: '#/components/responses/TrustedPublicationError' }
        '500': { $ref: '#/components/responses/TrustedPublicationError' }
  /v1/providers/{providerId}/endpoint-bindings/{bindingId}/challenges:
    post:
      operationId: createVerificationChallenge
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/providerId' }
        - { $ref: '#/components/parameters/bindingId' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/VerificationChallenge' } } } }
        '400': { $ref: '#/components/responses/TrustedPublicationError' }
        '401': { $ref: '#/components/responses/TrustedPublicationError' }
        '403': { $ref: '#/components/responses/TrustedPublicationError' }
        '404': { $ref: '#/components/responses/TrustedPublicationError' }
        '409': { $ref: '#/components/responses/TrustedPublicationError' }
        '503': { $ref: '#/components/responses/TrustedPublicationError' }
        '500': { $ref: '#/components/responses/TrustedPublicationError' }
  /v1/providers/{providerId}/endpoint-bindings/{bindingId}/challenges/{challengeId}/complete:
    post:
      operationId: completeVerificationChallenge
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/providerId' }
        - { $ref: '#/components/parameters/bindingId' }
        - { $ref: '#/components/parameters/challengeId' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/EndpointBinding' } } } }
        '400': { $ref: '#/components/responses/TrustedPublicationError' }
        '401': { $ref: '#/components/responses/TrustedPublicationError' }
        '403': { $ref: '#/components/responses/TrustedPublicationError' }
        '404': { $ref: '#/components/responses/TrustedPublicationError' }
        '409': { $ref: '#/components/responses/TrustedPublicationError' }
        '503': { $ref: '#/components/responses/TrustedPublicationError' }
        '500': { $ref: '#/components/responses/TrustedPublicationError' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  parameters:
    providerId: { name: providerId, in: path, required: true, schema: { $ref: '../schemas/trusted-publication.v1.schema.json#/$defs/identifier' } }
    agentId: { name: agentId, in: path, required: true, schema: { $ref: '../schemas/trusted-publication.v1.schema.json#/$defs/identifier' } }
    bindingId: { name: bindingId, in: path, required: true, schema: { $ref: '../schemas/trusted-publication.v1.schema.json#/$defs/identifier' } }
    challengeId: { name: challengeId, in: path, required: true, schema: { $ref: '../schemas/trusted-publication.v1.schema.json#/$defs/identifier' } }
    releaseId: { name: releaseId, in: path, required: true, schema: { $ref: '../schemas/trusted-publication.v1.schema.json#/$defs/identifier' } }
  schemas:
    CreateEndpointBindingRequest:
      type: object
      additionalProperties: false
      required: [endpoint, method, version]
      properties:
        endpoint: { type: string, format: uri }
        method: { type: string, enum: [http_well_known] }
        version: { $ref: '../schemas/trusted-publication.v1.schema.json#/$defs/semver' }
    EndpointBinding: { $ref: '../schemas/trusted-publication.v1.schema.json#/$defs/endpointBinding' }
    VerificationChallenge: { $ref: '../schemas/trusted-publication.v1.schema.json#/$defs/challenge' }
    AgentRelease: { $ref: '../schemas/trusted-publication.v1.schema.json#/$defs/agentRelease' }
    CreateAgentReleaseRequest:
      type: object
      additionalProperties: false
      required: [version, endpointBindingId]
      properties:
        version: { $ref: '../schemas/trusted-publication.v1.schema.json#/$defs/semver' }
        endpointBindingId: { $ref: '../schemas/trusted-publication.v1.schema.json#/$defs/identifier' }
    TrustedPublicationError: { $ref: '../schemas/trusted-publication.v1.schema.json#/$defs/error' }
  responses:
    TrustedPublicationError:
      description: Trusted publication error
      headers:
        x-nek-trace-id: { schema: { type: string } }
      content:
        application/json: { schema: { $ref: '#/components/schemas/TrustedPublicationError' } }